feat(#113): rename some classes
