+++ src/main/java/org/eolang/jeo/improvement/ImprovementEoFootprint.java
@@ -23,71 +23,70 @@
  */
 package org.eolang.jeo.improvement;
 
-import com.jcabi.log.Logger;
+import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 import java.util.Collection;
-import java.util.Collections;
+import java.util.stream.Collectors;
+import org.eolang.jeo.EoDefaultDirectory;
 import org.eolang.jeo.Improvement;
 import org.eolang.jeo.Representation;
+import org.eolang.jeo.representation.EoRepresentation;
 
 /**
- * Footprint of the representations as bytecode classes.
+ * Footprint of the EO's.
  *
  * @since 0.1.0
  */
-public final class BytecodeFootprint implements Improvement {
+public final class ImprovementEoFootprint implements Improvement {
 
     /**
-     * Where to save the bytecode classes.
+     * Where to save the EO.
      */
-    private final Path classes;
+    private final Path target;
 
     /**
      * Constructor.
-     * @param target Where to save the bytecode classes.
+     * @param home Where to save the EO.
      */
-    public BytecodeFootprint(final Path target) {
-        this.classes = target;
+    public ImprovementEoFootprint(final Path home) {
+        this.target = home;
     }
 
     @Override
-    public Collection<? extends Representation> apply(
+    public Collection<Representation> apply(
         final Collection<? extends Representation> representations
     ) {
-        representations.forEach(this::recompile);
-        return Collections.unmodifiableCollection(representations);
+        return representations.stream()
+            .map(Representation::toEO)
+            .map(EoRepresentation::new)
+            .peek(this::tryToSave)
+            .collect(Collectors.toList());
     }
 
     /**
-     * Recompile the Intermediate Representation.
-     * @param representation Intermediate Representation to recompile.
+     * Try to save XML to the target folder.
+     * @param representation XML to save.
      */
-    private void recompile(final Representation representation) {
+    private void tryToSave(final Representation representation) {
         final String name = representation.name();
+        final Path path = this.target.resolve(new EoDefaultDirectory().toPath())
+            .resolve(String.format("%s.xmir", name.replace('.', File.separatorChar)));
         try {
-            final byte[] bytecode = representation.toBytecode().asBytes();
-            final String[] subpath = name.split("\\.");
-            subpath[subpath.length - 1] = String.format("%s.class", subpath[subpath.length - 1]);
-            final Path path = Paths.get(this.classes.toString(), subpath);
-            Logger.info(
-                this,
-                "Recompiling '%s', bytecode instance '%s', bytes to save '%s'",
-                path,
-                representation.getClass(),
-                bytecode.length
-            );
             Files.createDirectories(path.getParent());
-            Files.write(path, bytecode);
-            Logger.info(
-                this,
-                "%s was recompiled successfully.",
-                path.getFileName().toString()
+            Files.write(
+                path,
+                representation.toEO().toString().getBytes(StandardCharsets.UTF_8),
+                StandardOpenOption.CREATE_NEW
             );
         } catch (final IOException exception) {
-            throw new IllegalStateException(String.format("Can't recompile '%s'", name), exception);
+            throw new IllegalStateException(
+                String.format("Can't save XML to %s", path),
+                exception
+            );
         }
     }
 }
