+++ src/test/java/org/eolang/jeo/representation/BytecodeTransformationTest.java
@@ -21,40 +21,45 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.improvement;
+package org.eolang.jeo.representation;
 
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collections;
-import org.eolang.jeo.representation.Eo;
-import org.eolang.jeo.representation.EoRepresentation;
+import org.cactoos.bytes.BytesOf;
+import org.cactoos.bytes.UncheckedBytes;
+import org.cactoos.io.ResourceOf;
 import org.hamcrest.MatcherAssert;
-import org.hamcrest.io.FileMatchers;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 /**
- * Test case for {@link BytecodeFootprint}.
+ * Tests for {@link BytecodeTransformation}.
  *
  * @since 0.1.0
  */
-class BytecodeFootprintTest {
+class BytecodeTransformationTest {
 
     @Test
-    void appliesSuccessfully(@TempDir final Path temp) {
-        final String expected = "jeo.xmir.Fake";
-        new BytecodeFootprint(temp).apply(
-            Collections.singleton(new EoRepresentation(new Eo(expected)))
+    void transpilesSuccessfully(@TempDir final Path temp) throws IOException {
+        final String name = "MethodByte.class";
+        Files.write(
+            temp.resolve(name),
+            new UncheckedBytes(new BytesOf(new ResourceOf(name))).asBytes()
         );
+        new BytecodeTransformation(temp, temp).transpile();
         MatcherAssert.assertThat(
-            String.format(
-                "Bytecode file was not saved for the representation with the name '%s'",
-                expected
+            String.format("Can't find the transpiled file for the class '%s'.", name),
+            Files.exists(
+                temp.resolve("jeo")
+                    .resolve("xmir")
+                    .resolve("org")
+                    .resolve("eolang")
+                    .resolve("jeo")
+                    .resolve("MethodByte.xmir")
             ),
-            temp.resolve("jeo")
-                .resolve("xmir")
-                .resolve("Fake.class")
-                .toFile(),
-            FileMatchers.anExistingFile()
+            Matchers.is(true)
         );
     }
 }
